Fix compile errors

+++ src/game/Path.h
@@ -98,7 +98,7 @@ class Path
                 float xd = node.x - prev.x;
                 float yd = node.y - prev.y;
                 float zd = node.z - prev.z;
-                len += std::sqrt(xd*xd + yd*yd + zd*zd);
+                len += sqrtf(xd*xd + yd*yd + zd*zd);
             }
             return len;
         }
@@ -113,7 +113,7 @@ class Path
                 float xd = x - node.x;
                 float yd = y - node.y;
                 float zd = z - node.z;
-                len += std::sqrt(xd*xd + yd*yd + zd*zd);
+                len += sqrtf(xd*xd + yd*yd + zd*zd);
             }
 
             return len;
